+++ arquillian-container-liferay/src/main/java/org/arquillian/container/liferay/remote/enricher/Inject.java
@@ -12,21 +12,14 @@
  * details.
  */
 
-package com.liferay.arquillian.container.enricher;
+package org.arquillian.container.liferay.remote.enricher;
 
-import org.jboss.arquillian.container.test.spi.RemoteLoadableExtension;
-import org.jboss.arquillian.test.spi.TestEnricher;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
 
 /**
  * @author Carlos Sierra Andrés
  */
-public class LiferayEnricherRemoteExtension implements RemoteLoadableExtension {
-
-	@Override
-	public void register(ExtensionBuilder builder) {
-		if (Validate.classExists("org.osgi.framework.Bundle")) {
-			builder.service(TestEnricher.class, LiferayTestEnricher.class);
-		}
-	}
-
+@Retention(RetentionPolicy.RUNTIME)
+public @interface Inject {
 }
