+++ arquillian-container-liferay/src/main/resources/META-INF/services/org.jboss.arquillian.container.test.spi.TestRunner
@@ -1,1 +1,1 @@
-com.liferay.arquillian.container.runner.JUnitBundleTestRunner
+org.arquillian.container.liferay.remote.runner.JUnitBundleTestRunner
